AK3: Only perform super device size checks when necessary
For logical partitions, the logic of AnyKernel3 is: as long as the image size is
inconsistent with the partition size, this partition will be re-created; if the
re-creation of the partition fails, the existing partition will be resized.

For the vendor_dlkm image using erofs, its size is always inconsistent with the
previous one after each regeneration. Therefore, in order to avoid having to
re-create the vendor_dlkm partition for each installation, a simple method is to
fill the regenerated vendor_dlkm image to the same size as the vendor_dlkm partition.

For vendor_dlkm images using ext4, we only perform super device size checks when
expanding the image. Because more and more rom maintainers are also willing to
reserve space for the vendor_dlkm partition (although this is not necessary).