Introducing a rollback mechanism
Unlike other kernels, Melt Kernel will modify and flash the boot
partition, vendor_dlkm partition, and vendor_boot partition in
sequence during the installation process.

In special cases, the boot partition and vendor_dlkm partition may
be modified and flashed, but the vendor_boot partition may fail.
The best result would be to make the device unbootable so that the
user is aware of the problem. However, this is not the case.
Sometimes the device is still able to boot, and because the boot
partition has been successfully flashed, the kernel version strings
has changed, giving the user the illusion that the kernel has been
successfully installed.

Therefore, we designed this rollback mechanism to install the
unmodified boot image and vendor_dlkm image to the device if the
installation process fails.

For now, we will not consider the situation where the boot partition
has been successfully flashed but the vendor_dlkm partition has not
been successfully processed, because Melt has already handled the
entire installation process perfectly enough.